--- client/src/pages/actions/Edit.tsx
@@ -1,7 +1,0 @@
-import type { IResourceComponentsProps } from '@refinedev/core'
-import { MuiEditInferencer } from '@refinedev/inferencer/mui'
-import type { FC } from 'react'
-
-export const ActionEdit: FC<IResourceComponentsProps> = () => (
-  <MuiEditInferencer />
-)
