+++ client/src/pages/actions/index.ts
@@ -1,3 +1,2 @@
-export { ActionEdit } from './Edit'
 export { ActionList } from './List'
 export { ActionShow } from './Show'
