+++ client/src/pages/flow/Show.tsx
@@ -2,7 +2,6 @@ import { AlertColor } from '@mui/material/Alert/Alert'
 import Grid from '@mui/material/Grid'
 import { Box } from '@mui/system'
 import { GetListResponse, useList } from '@refinedev/core'
-import { KBarProvider, KBarProviderProps, RefineKbar } from '@refinedev/kbar'
 import { type FC, useEffect, useState } from 'react'
 
 import { components } from '../../../openapi'
@@ -14,7 +13,7 @@ import {
   SidebarTreeItemClickStatesProvider,
   SidebarTreeItemMouseStatesProvider,
 } from '../../context/SidebarTreeItemStatesContext'
-import { useAction, useActionDispatch } from '../../hooks/ActionHooks'
+import { useAction } from '../../hooks/ActionHooks'
 import { checkIfDuplicatedActions } from '../../utils/helpers'
 
 export const FlowShow: FC = () => {
@@ -32,12 +31,6 @@ export const FlowShow: FC = () => {
         type?: AlertColor
       }
   >(false)
-
-  const dispatch = useActionDispatch()
-
-  const [kBarActions, setKbarActions] = useState<KBarProviderProps['actions']>(
-    []
-  )
 
   useEffect(() => {
     setRenderEndSidebar(nodeId !== '')
@@ -80,23 +73,8 @@ export const FlowShow: FC = () => {
           setData(actions)
         }
       }
-
-      const kbar: KBarProviderProps['actions'] = []
-
-      actions?.data.forEach((action) => {
-        kbar.push({
-          id: action.id,
-          name: `${action.title} (${action.id})`,
-          perform: () =>
-            dispatch?.({
-              type: 'set-active-action',
-              id: action.id,
-            }),
-        })
-      })
-      setKbarActions(kbar)
     }
-  }, [actions, dataReceived, dispatch])
+  }, [actions, dataReceived])
 
   return (
     <SidebarTreeItemMouseStatesProvider>
@@ -109,27 +87,24 @@ export const FlowShow: FC = () => {
           {alert && typeof alert !== 'boolean' && <AlertBanner data={alert} />}
           {dataReceived && (
             <>
-              <KBarProvider actions={kBarActions}>
-                <RefineKbar />
-                <Grid item xs={7} sx={{ height: 'calc(100vh - 68px)' }}>
-                  <SidebarFlow actions={dataReceived} />
-                </Grid>
-                <Grid item xs={29} sx={{ height: 'calc(100vh - 68px)' }}>
-                  <ActionsFlow actions={dataReceived} />
-                </Grid>
-                {renderEndSidebar && (
-                  <Box
-                    sx={{
-                      height: 'calc(100vh - 68px)',
-                      position: 'fixed',
-                      right: 0,
-                      top: 68,
-                    }}
-                  >
-                    <SecondSidebarFlow actions={dataReceived} nodeId={nodeId} />
-                  </Box>
-                )}
-              </KBarProvider>
+              <Grid item xs={7} sx={{ height: 'calc(100vh - 68px)' }}>
+                <SidebarFlow actions={dataReceived} />
+              </Grid>
+              <Grid item xs={29} sx={{ height: 'calc(100vh - 68px)' }}>
+                <ActionsFlow actions={dataReceived} />
+              </Grid>
+              {renderEndSidebar && (
+                <Box
+                  sx={{
+                    height: 'calc(100vh - 68px)',
+                    position: 'fixed',
+                    right: 0,
+                    top: 68,
+                  }}
+                >
+                  <SecondSidebarFlow actions={dataReceived} nodeId={nodeId} />
+                </Box>
+              )}
             </>
           )}
         </Grid>
